User commit

+++ src/main/java/com/example/demo/dto/CapacityAnalysisResultDto.java
@@ -1,4 +1,3 @@
-// File: src/main/java/com/example/demo/dto/CapacityAnalysisResultDto.java
 package com.example.demo.dto;
 
 import java.time.LocalDate;
@@ -6,13 +5,12 @@ import java.util.Map;
 
 public class CapacityAnalysisResultDto {
     private boolean risky;
-    private Map<LocalDate, Integer> capacityByDate;
-    
+    private Map<LocalDate, Double> capacityByDate;
+
+    // Getters and setters
     public boolean isRisky() { return risky; }
     public void setRisky(boolean risky) { this.risky = risky; }
     
-    public Map<LocalDate, Integer> getCapacityByDate() { return capacityByDate; }
-    public void setCapacityByDate(Map<LocalDate, Integer> capacityByDate) { 
-        this.capacityByDate = capacityByDate; 
-    }
+    public Map<LocalDate, Double> getCapacityByDate() { return capacityByDate; }
+    public void setCapacityByDate(Map<LocalDate, Double> capacityByDate) { this.capacityByDate = capacityByDate; }
 }
